fix(editor): Fix type safety, memory leak, and extract canvas preview composable
- Remove `as never` cast in executionStarted handler by adding missing
  IExecutionBase/IWorkflowDb fields and using createRunExecutionData()
- Fix threadPushRef memory leak by cleaning up entries in both finally
  blocks when runs complete
- Fix useI18n() called at module scope in module.descriptor.ts — use
  direct i18n singleton import instead
- Add console.warn to silent return in updateNodeExecutionStatus
- Extract canvas preview state into useCanvasPreview composable,
  reducing InstanceAiView.vue by ~175 lines
- Fix DataTableTable indentation and remove unused .openLink CSS
- Preserve ChatHub markdown appearance after ChatMarkdownChunk switched
  to inherit — set explicit font-size/line-height/color on ChatMessage
  and ChatArtifactViewer wrappers
- Fix pre-existing test expecting removed "Open in editor" text
