+++ packages/frontend/editor-ui/src/features/ai/chatHub/components/ChatArtifactViewer.vue
@@ -135,6 +135,11 @@ const markdownContent = computed(() => ({
 .content {
 	flex: 1;
 	overflow-x: hidden;
+	// ChatMarkdownChunk uses `inherit` — set the values previously hardcoded
+	// in the chunk so the artifact viewer preserves existing appearance.
+	color: var(--color--text--shade-1);
+	font-size: var(--font-size--md);
+	line-height: var(--line-height--xl);
 }
 
 .iframe {
